% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hello.R
\name{say_hello}
\alias{say_hello}
\title{say_hello A function to say hello}
\usage{
say_hello(someone = "You")
}
\arguments{
\item{someone}{the name of the person to say hello to (type = character)}
}
\value{
A message with Hello to someone giving their name
}
\description{
say_hello A function to say hello
}
\examples{
say_hello()
say_hello(someone = "Katerina")
}
